requirements.txt
appdirs==1.4.3
Babel==2.5.3
blinker==1.4
click==6.7
Flask==0.12
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Navigation==0.2.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==3.0.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
passlib==1.7.1
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2017.3
six==1.10.0
speaklater==1.3
SQLAlchemy==1.1.6
Werkzeug==0.12.1
WTForms==2.1
loremipsum-1.0.5
psycopg2-2.7.3.2
sortedcontainers==1.5.9
mock-2.0.0
Flask-Testing-0.7.1
